requirements.txt
pillow==8.3.1
numpy==1.21.0
torch==1.10.0
torchvision==0.11.1